Use a full path to the current ruby for subshell
If we rely on the system PATH for the ruby to use in the subshell, and
the user isn't using RVM or similar, then the subshell may use a ruby
that doesn't contain turbolib dependencies.

Specifically, this has been a problem with packaged OSX rubies: running
`turbot` uses the ruby we supply in the package, but when the runner
spawns a subshell to run a ruby scraper, the system Ruby (1.8.7 on older
OSX) is used, which (a) lacks the turbot-ruby-gems and (b) segfaults as
a result.