Pass `retrieved_at` to validator.
Previously we were omitting this. We're not sure why, but it seems
likely it was because simple transformers included a retrieved_at field
which was not validating -- perhaps ones created prior to us extending
json-schema with our own `date` format.

See PR at https://github.com/openc/openc-schema/pull/52 which adds
optional `retrieved_at` to all schemas, and tests simple subsidiaries
are valid.